solve 2302 - 극장 좌석
silver 1
기초 dp

+++ Algorithm/Algorithm/6198 - 옥상 정원 꾸미기.cpp
@@ -8,8 +8,8 @@
 #define endl "\n"
 using namespace std;
 
-int n, num, ans;
-stack<int> s;
+ll n, num, ans;
+stack<ll> s;
 
 int main()
 {
